requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==2.2.5
django-crispy-forms==1.8.0
django-filter==2.2.0
django-tables2==2.2.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pkg-resources==0.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
